// apps/fusion/project/ws/app/src/lib/routes/infy/routes/navigator/models/navigator.model.ts
export interface IFsPlayground {
  playground_name: string
  playground_id: number
  playground_desc: string
  playground_link: string
  playground_image: string
}

export interface ILpInternalCertification {
  virtual_proctoring: boolean
  virtual_proctoring_id: string
  lp_internal_certification_link: string
  alternate_parent_id?: string[]
  lp_internal_certification_name: string
  lp_internal_certification_description: string
  is_alternate: boolean
  lp_internal_certification_id: string
}

export interface IFsInternalCertification {
  internal_certification_id: string
  internal_certification_name: string
  internal_certification_desc: string
  internal_certification_link: string
  internal_certification_image: string
}

export interface IExternalCertification {
  virtual_proctoring: boolean
  virtual_proctoring_id?: string
  alternate_parent_id: string[]
  lp_external_certification_name: string
  lp_external_certification_id: string
  lp_external_certification_link: string
  is_alternate: boolean
  lp_external_certification_description: string
}

export interface IFsExternalCertification {
  external_certification_id: number
  external_certification_name: string
  external_certification_desc: string
  external_certification_image: string
  external_certification_link: string
}
// ...
export interface IFsData {
  fs_id: number
  fs_name: string
  fs_desc: string
  fs_image: string
  fs_linked_program: string
  fs_internal_certification: IFsInternalCertification[]
  fs_playground: IFsPlayground[]
  fs_external_certification: IFsExternalCertification[]
  fs_course: IFsCourse[]
}

export interface ICourse {
  course_id: number
  course_name: string
  course_description: string
  course_link: string
  course_image: string
}

export interface IFsCourse {
  course_id: number
  course_name: string
  course_desc: string
  course_lex_link: string
  course_image: string
  course_time: number
}
// ...
export interface IFsCardModel {
  thumbnail: string
  title: string
  description: string
  routeButton: string
  type: string
}